% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appRiori.R
\name{appRiori}
\alias{appRiori}
\title{Running appRiori}
\usage{
appRiori(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\value{
No return value, called for side effects
}
\description{
This function runs the appRiori Shiny app.
}
\examples{

\dontrun{
# Run the app with default settings
library(appRiori)
appRiori()
}


}
\seealso{
\code{\link[shiny:runApp]{runApp()}}
}
