% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{contrasts_summary}
\alias{contrasts_summary}
\title{Summary for hypr contrasts}
\usage{
contrasts_summary(model)
}
\arguments{
\item{model}{a model of class \code{\link[stats::lm]{stats::lm}}, \code{\link[stats::lm]{stats::glm}}
or \code{\link[lme4::lmer]{lmer::lmer}}}
}
\value{
A results table of class \code{hypr_contrasts_results}.
}
\description{
Extract the estimates and the inferential tests for a contrasts
defined by \code{\link[hypr::hyper]{hypr::hyper}}
}
\examples{

\dontrun{
}
}
