% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.hypr_contrasts_results}
\alias{print.hypr_contrasts_results}
\title{Print for hypr contrasts}
\usage{
\method{print}{hypr_contrasts_results}(x, ...)
}
\arguments{
\item{x}{an object of class \code{hypr_contrasts_results}}

\item{...}{additional options (not used at the moment)}
}
\value{
A results table
}
\description{
Pretty print the estimates and the inferential tests for a contrasts
defined by \code{\link[hypr]{hypr}}
}
